+++ GeometricFlux.jl/test/layers/conv.jl
@@ -3,10 +3,15 @@ using Flux: Dense
 in_channel = 3
 out_channel = 5
 N = 4
-adj = [0. 1. 0. 1.;
+adj = T[0. 1. 0. 1.;
        1. 0. 1. 0.;
        0. 1. 0. 1.;
        1. 0. 1. 0.]
+
+adj_single_vertex =   T[0. 0. 0. 1.;
+                        0. 0. 0. 0.;
+                        0. 0. 0. 1.;
+                        1. 0. 1. 0.]
 
 
 @testset "layer" begin
@@ -21,8 +26,17 @@ adj = [0. 1. 0. 1.;
             Y = gc(X)
             @test size(Y) == (out_channel, N)
             
-            g = Zygote.gradient(() -> sum(gc(X)), params(gc))
-            @test length(g.grads) == 2
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            Y = gc(transpose(Xt))
+            @test size(Y) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(gc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), gc)[1]
+            @test size(g.weight) == size(gc.weight)
+            @test size(g.bias) == size(gc.bias)
         end
 
         @testset "layer without graph" begin
@@ -36,8 +50,18 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError gc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(gc(fg))), params(gc))
-            @test length(g.grads) == 4
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            fgt = FeaturedGraph(adj, transpose(Xt))
+            fgt_ = gc(fgt)
+            @test size(node_feature(fgt_)) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(node_feature(gc(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), gc)[1]
+            @test size(g.weight) == size(gc.weight)
+            @test size(g.bias) == size(gc.bias)
         end
     end
 
@@ -57,8 +81,17 @@ adj = [0. 1. 0. 1.;
             Y = cc(X)
             @test size(Y) == (out_channel, N)
 
-            g = Zygote.gradient(() -> sum(cc(X)), params(cc))
-            @test length(g.grads) == 3
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            Y = cc(transpose(Xt))
+            @test size(Y) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(cc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), cc)[1]
+            @test size(g.weight) == size(cc.weight)
+            @test size(g.bias) == size(cc.bias)
         end
 
         @testset "layer without graph" begin
@@ -75,8 +108,18 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError cc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(cc(fg))), params(cc))
-            @test length(g.grads) == 4
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            fgt = FeaturedGraph(adj, transpose(Xt))
+            fgt_ = cc(fgt)
+            @test size(node_feature(fgt_)) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(node_feature(cc(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), cc)[1]
+            @test size(g.weight) == size(cc.weight)
+            @test size(g.bias) == size(cc.bias)
         end
     end
 
@@ -92,9 +135,20 @@ adj = [0. 1. 0. 1.;
             Y = gc(X)
             @test size(Y) == (out_channel, N)
 
-            # Test that the gradient can be computed
-            g = Zygote.gradient(() -> sum(gc(X)), params(gc))
-            @test length(g.grads) == 5
+            
+
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            Y = gc(transpose(Xt))
+            @test size(Y) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(gc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), gc)[1]
+            @test size(g.weight1) == size(gc.weight1)
+            @test size(g.weight2) == size(gc.weight2)
+            @test size(g.bias) == size(gc.bias)
         end
 
         @testset "layer without graph" begin
@@ -110,82 +164,151 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError gc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(gc(fg))), params(gc))
-            @test length(g.grads) == 5
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            fgt = FeaturedGraph(adj, transpose(Xt))
+            fgt_ = gc(fgt)
+            @test size(node_feature(fgt_)) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(node_feature(gc(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), gc)[1]
+            @test size(g.weight1) == size(gc.weight1)
+            @test size(g.weight2) == size(gc.weight2)
+            @test size(g.bias) == size(gc.bias)
         end
     end
 
     @testset "GATConv" begin
+
         X = rand(in_channel, N)
 
-        @testset "layer with graph" begin
-            @testset "concat=true" begin
-                for heads = [1, 6]
-                    gat = GATConv(adj, in_channel=>out_channel, heads=heads, concat=true)
-                    @test graph(gat.fg) == [[2,4], [1,3], [2,4], [1,3]]
-                    @test size(gat.weight) == (out_channel * heads, in_channel)
-                    @test size(gat.bias) == (out_channel * heads,)
-                    @test size(gat.a) == (2*out_channel, heads, 1)
+        for adj_gat in [adj, adj_single_vertex]
 
-                    Y = gat(X)
-                    @test size(Y) == (out_channel * heads, N)
+            @testset "layer with graph" begin
+                @testset "concat=true" begin
+                    for heads = [1, 6]
+                        gat = GATConv(adj_gat, in_channel=>out_channel, heads=heads, concat=true)
 
-                    g = Zygote.gradient(() -> sum(gat(X)), params(gat))
-                    @test length(g.grads) == 5
+                        if adj_gat == adj
+                            @test graph(gat.fg) == [[2,4], [1,3], [2,4], [1,3]]
+                        end
+                        if adj_gat == adj_single_vertex
+                            @test graph(gat.fg) == [[4], Int64[], [4], [1, 3]]
+                        end
+
+                        @test size(gat.weight) == (out_channel * heads, in_channel)
+                        @test size(gat.bias) == (out_channel * heads,)
+                        @test size(gat.a) == (2*out_channel, heads)
+
+                        Y = gat(X)
+                        @test size(Y) == (out_channel * heads, N)
+
+                        # Test with transposed features
+                        Xt = rand(N, in_channel)
+                        Y = gat(transpose(Xt))
+                        @test size(Y) == (out_channel * heads, N)
+                            
+                        g = Zygote.gradient(x -> sum(gat(x)), X)[1]
+                        @test size(g) == size(X)
+
+                        g = Zygote.gradient(model -> sum(model(X)), gat)[1]
+                        @test size(g.weight) == size(gat.weight)
+                        @test size(g.bias) == size(gat.bias)
+                        @test size(g.a) == size(gat.a)
+                    end
+                end
+
+                @testset "concat=false" begin
+                    for heads = [1, 6]
+                        gat = GATConv(adj_gat, in_channel=>out_channel, heads=heads, concat=false)
+                        if adj_gat == adj
+                            @test graph(gat.fg) == [[2,4], [1,3], [2,4], [1,3]]
+                        end
+                        if adj_gat == adj_single_vertex
+                            @test graph(gat.fg) == [[4], Int64[], [4], [1, 3]]
+                        end
+                        @test size(gat.weight) == (out_channel * heads, in_channel)
+                        @test size(gat.bias) == (out_channel * heads,)
+                        @test size(gat.a) == (2*out_channel, heads)
+
+                        Y = gat(X)
+                        @test size(Y) == (out_channel, N)
+
+                        # Test with transposed features
+                        Xt = rand(N, in_channel)
+                        Y = gat(transpose(Xt))
+                        @test size(Y) == (out_channel, N)
+
+                        g = Zygote.gradient(x -> sum(gat(x)), X)[1]
+                        @test size(g) == size(X)
+
+                        g = Zygote.gradient(model -> sum(model(X)), gat)[1]
+                        @test size(g.weight) == size(gat.weight)
+                        @test size(g.bias) == size(gat.bias)
+                        @test size(g.a) == size(gat.a)
+                    end
                 end
             end
 
-            @testset "concat=false" begin
-                for heads = [1, 6]
-                    gat = GATConv(adj, in_channel=>out_channel, heads=heads, concat=false)
-                    @test graph(gat.fg) == [[2,4], [1,3], [2,4], [1,3]]
-                    @test size(gat.weight) == (out_channel * heads, in_channel)
-                    @test size(gat.bias) == (out_channel * heads,)
-                    @test size(gat.a) == (2*out_channel, heads, 1)
+            @testset "layer without graph" begin
+                fg = FeaturedGraph(adj_gat, X)
+                
+                @testset "concat=true" begin
+                    for heads = [1, 6]
+                        gat = GATConv(in_channel=>out_channel, heads=heads, concat=true)
+                        @test size(gat.weight) == (out_channel * heads, in_channel)
+                        @test size(gat.bias) == (out_channel * heads,)
+                        @test size(gat.a) == (2*out_channel, heads)
 
-                    Y = gat(X)
-                    @test size(Y) == (out_channel * heads, 1)
+                        fg_ = gat(fg)
+                        Y = node_feature(fg_)
+                        @test size(Y) == (out_channel * heads, N)
+                        @test_throws AssertionError gat(X)
 
-                    g = Zygote.gradient(() -> sum(gat(X)), params(gat))
-                    @test length(g.grads) == 5
+                        # Test with transposed features
+                        Xt = rand(N, in_channel)
+                        fgt = FeaturedGraph(adj_gat, transpose(Xt))
+                        fgt_ = gat(fgt)
+                        @test size(node_feature(fgt_)) == (out_channel * heads, N)
+
+                        g = Zygote.gradient(x -> sum(node_feature(gat(x))), fg)[1]
+                        @test size(g[].nf) == size(X)
+
+                        g = Zygote.gradient(model -> sum(node_feature(model(fg))), gat)[1]
+                        @test size(g.weight) == size(gat.weight)
+                        @test size(g.bias) == size(gat.bias)
+                        @test size(g.a) == size(gat.a)
+                    end
                 end
-            end
-        end
 
-        @testset "layer without graph" begin
-            fg = FeaturedGraph(adj, X)
-            
-            @testset "concat=true" begin
-                for heads = [1, 6]
-                    gat = GATConv(in_channel=>out_channel, heads=heads, concat=true)
-                    @test size(gat.weight) == (out_channel * heads, in_channel)
-                    @test size(gat.bias) == (out_channel * heads,)
-                    @test size(gat.a) == (2*out_channel, heads, 1)
+                @testset "concat=false" begin
+                    for heads = [1, 6]
+                        gat = GATConv(in_channel=>out_channel, heads=heads, concat=false)
+                        @test size(gat.weight) == (out_channel * heads, in_channel)
+                        @test size(gat.bias) == (out_channel * heads,)
+                        @test size(gat.a) == (2*out_channel, heads)
 
-                    fg_ = gat(fg)
-                    Y = node_feature(fg_)
-                    @test size(Y) == (out_channel * heads, N)
-                    @test_throws AssertionError gat(X)
+                        fg_ = gat(fg)
+                        Y = node_feature(fg_)
+                        @test size(Y) == (out_channel, N)
+                        @test_throws AssertionError gat(X)
 
-                    g = Zygote.gradient(() -> sum(node_feature(gat(fg))), params(gat))
-                    @test length(g.grads) == 5
-                end
-            end
+                        # Test with transposed features
+                        Xt = rand(N, in_channel)
+                        fgt = FeaturedGraph(adj_gat, transpose(Xt))
+                        fgt_ = gat(fgt)
+                        @test size(node_feature(fgt_)) == (out_channel, N)
 
-            @testset "concat=false" begin
-                for heads = [1, 6]
-                    gat = GATConv(in_channel=>out_channel, heads=heads, concat=false)
-                    @test size(gat.weight) == (out_channel * heads, in_channel)
-                    @test size(gat.bias) == (out_channel * heads,)
-                    @test size(gat.a) == (2*out_channel, heads, 1)
+                        g = Zygote.gradient(x -> sum(node_feature(gat(x))), fg)[1]
+                        @test size(g[].nf) == size(X)
 
-                    fg_ = gat(fg)
-                    Y = node_feature(fg_)
-                    @test size(Y) == (out_channel * heads, 1)
-                    @test_throws AssertionError gat(X)
-
-                    g = Zygote.gradient(() -> sum(node_feature(gat(fg))), params(gat))
-                    @test length(g.grads) == 5
+                        g = Zygote.gradient(model -> sum(node_feature(model(fg))), gat)[1]
+                        @test size(g.weight) == size(gat.weight)
+                        @test size(g.bias) == size(gat.bias)
+                        @test size(g.a) == size(gat.a)
+                    end
                 end
             end
         end
@@ -202,8 +325,17 @@ adj = [0. 1. 0. 1.;
             Y = ggc(X)
             @test size(Y) == (out_channel, N)
 
-            g = Zygote.gradient(() -> sum(ggc(X)), params(ggc))
-            @test length(g.grads) == 13
+            
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            Y = ggc(transpose(Xt))
+            @test size(Y) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(ggc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), ggc)[1]
+            @test size(g.weight) == size(ggc.weight)
         end
 
         @testset "layer without graph" begin
@@ -216,8 +348,18 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError ggc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(ggc(fg))), params(ggc))
-            @test length(g.grads) == 13
+            
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            fgt = FeaturedGraph(adj, transpose(Xt))
+            fgt_ = ggc(fgt)
+            @test size(node_feature(fgt_)) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(node_feature(ggc(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), ggc)[1]
+            @test size(g.weight) == size(ggc.weight)
         end
     end
 
@@ -230,8 +372,18 @@ adj = [0. 1. 0. 1.;
             Y = ec(X)
             @test size(Y) == (out_channel, N)
 
-            g = Zygote.gradient(() -> sum(ec(X)), params(ec))
-            @test length(g.grads) == 4
+            
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            Y = ec(transpose(Xt))
+            @test size(Y) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(ec(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), ec)[1]
+            @test size(g.nn.W) == size(ec.nn.W)
+            @test size(g.nn.b) == size(ec.nn.b)
         end
 
         @testset "layer without graph" begin
@@ -243,8 +395,19 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError ec(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(ec(fg))), params(ec))
-            @test length(g.grads) == 4
+            
+            # Test with transposed features
+            Xt = rand(N, in_channel)
+            fgt = FeaturedGraph(adj, transpose(Xt))
+            fgt_ = ec(fgt)
+            @test size(node_feature(fgt_)) == (out_channel, N)
+            
+            g = Zygote.gradient(x -> sum(node_feature(ec(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), ec)[1]
+            @test size(g.nn.W) == size(ec.nn.W)
+            @test size(g.nn.b) == size(ec.nn.b)
         end
     end
 end