+++ cluster/cluster_1.jl
@@ -49,6 +49,7 @@ function distance_3()
     gallery = X
     query = X
     println("knn...")
+    brutetree = BruteTree(data)  # 暴力搜索树
     kdtree = KDTree(gallery, leafsize=4)   # 同index.add(gallery) 
     idxs, dists = knn(kdtree, query, k, true)  # 单线程的, 很慢.  # query查询. (70184,)
     dists = vcat((hcat(i...) for i in dists)...)  # 转换 shape
@@ -108,8 +109,8 @@ end
 # @time distance_2()
 @time distance_3()
 
+
 #=
-
-
+2020.1
 
 =#