+++ cluster/stream/test_2.jl
@@ -1,25 +1,5 @@
-using Dates, CSV, Plots, PlotThemes
-using ProgressMeter
-# Interact, HTTP, SingularSpectrumAnalysis, JuliaDB,
-using OnlineStats
-
-# plotly()
-# theme(:bright)
-ENV["GKSwstype"] = "100"
-ENV["GKS_ENCODING"]="utf8"
-gr()
+include("onlinestats_1.jl")
 
 
-o = Partition(Series(Mean(), Extrema()))
-y = randn()
-
-@showprogress for _ in 1:10^3
-    fit!(o,  global y += randn())
-end
-
-# println(o)
-# default(show=true)
-plot(o)   
-# display()
-
+println("aaa")